requirements.txt
# Core
openai
python-dotenv

# Audio transcription
torch
git+https://github.com/openai/whisper.git


flask
flask-cors
